Add interface for Donation Accounts and refine `ImportedWalletAccountPayload`.
