+++ src/common/data/models/AccountPayload/HexaAccountPayloads.ts
@@ -2,7 +2,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { iconForAccountKind } from '../../../../utils/accounts/IconUtils';
 import AccountKind from "../../enums/AccountKind";
 import BitcoinUnit from '../../enums/BitcoinUnit';
-import { HexaAccountPayload } from './AccountPayload';
+import { HexaAccountPayload } from './Interfaces';
 
 
 interface ConstructorProps {
@@ -15,7 +15,6 @@ interface ConstructorProps {
   customDescription?: string | null;
   secondaryAccountUUIDs?: string[];
 }
-
 
 export class TestAccountPayload implements HexaAccountPayload {
   uuid: string = uuidv4();
@@ -167,49 +166,6 @@ export class TrustedContactsAccountPayload implements HexaAccountPayload {
     secondaryAccountUUIDs,
   }: ConstructorProps = {}) {
     this.title = title || "Joint Account";
-    this.accountNumber = accountNumber || 0;
-    this.displayOrder = displayOrder || null;
-    this.balance = balance || 0;
-    this.unit = unit || BitcoinUnit.SATS;
-    this.customDisplayName = customDisplayName || null;
-    this.customDescription = customDescription || null;
-    this.secondaryAccountUUIDs = secondaryAccountUUIDs || [];
-  }
-
-  get imageSource(): NodeRequire {
-    return iconForAccountKind(this.kind);
-  }
-
-  get isPrimaryAccount(): boolean {
-    return this.secondaryAccountUUIDs.length === 0;
-  }
-}
-
-
-export class DonationAccountPayload implements HexaAccountPayload {
-  uuid: string = uuidv4();
-  title: string;
-  shortDescription: string = "Directly Accept Donations";
-  kind: AccountKind = AccountKind.DONATION;
-  accountNumber: number;
-  displayOrder: number | null;
-  balance: number;
-  unit: BitcoinUnit;
-  customDisplayName: string | null;
-  customDescription: string | null;
-  secondaryAccountUUIDs: string[];
-
-  constructor({
-    title,
-    accountNumber,
-    displayOrder,
-    balance,
-    unit,
-    customDisplayName,
-    customDescription,
-    secondaryAccountUUIDs,
-  }: ConstructorProps = {}) {
-    this.title = title || "Donation Account";
     this.accountNumber = accountNumber || 0;
     this.displayOrder = displayOrder || null;
     this.balance = balance || 0;
